Apps/Greeting_Card_Generator/requirements.txt
gradio
google-generativeai
